Added subsampling and supersampling
shape.py:

Added functions supersample and subsample, which do as they say. they directly apply the sampling we want and change the mesh (and the shape values) inside the shape to the new mesh.

preprocess.py:

Fixed logic regarding whether a shape needs pre-processing or not. Added pandas dataframe for future use.